Fix 12: Make image threshold algorithm configurable
Added bin-based ML thresholding algorithm. Added training node that examines
images and masks to generate the thresholded bin indices for the binning
algorithm. Also added a default set of bins to the binning algorithm, and
added camera data for the Logitech C270 USB camera. Added a launcher for the
cone detector for using a web cam.