Reduced logging and simplified message publication
Commented out some logging messages. Simplified message publication in
view_mission by separating the robot position and waypoint position
publishers, and by latching the waypoint messages. Clarified in the
navigator script that the local X axis corresponds to GPS East, always,
so that we do not need to handle rotation from the GPS frame to the map
frame.