Make turning constant more intuitive
Modified naming and use of the constant controlling the turn rate
while driving to a cone. Changed the name to use 'turning' instead
of 'heading'. Also modified its use and default value to mean that
the constant represents the number of seconds in which the heading
error will beome zero, based on the commanded turning.