Start to add support for manual control driving to cone
Added a parameter telling whether to use GUIDED or MANUAL control
for driving to the cone. Changed the proportional control to send
R/C commands if in MANUAL control. Still need to change state
machine to go into MANUAL instead of GUIDED mode if using manual
control, and to set R/C throttle to zero when exiting MANUAL mode.