Finish option to use manual mode when driving to cone
Modified state diagram to show that MANUAL mode may be used for
driving to cone, and showed transitions where manual speed should
be set to zero. Modified navigator node to move to manual mode,
if configured, when driving to cone, and to set manual speed
to zero when required by state transitions.