Add configuration for manual control, and fix sign of turning
Created a new YAML configuration for manual control of the rover
while approaching a cone. Fixed a bug in the navigator where the
sign of turning control was wrong while approaching the cone. At
the same time, modified the sense of set_manual_speed() so that
positive turning is to the left, just as for usual ROS cmd_vel
messages.