Collapse states to handle cones without intermediate waypoints
Modified the state machine to collapse FOLLOWING_WAYPOINTS and
EXPECTING_CONE, so that no intermediate waypoint is needed to
put the rover in a state where it is looking for a cone. Also
modified the mock rover configuration to switch the PWM values
for left and right steering, to match the configuration on
Victoria.