Change turning code to have a target heading
Modified navigator node to save away the target heading when
deciding to turn toward the next waypoint, in preparation for
using the same code for the recovery mode. Added configuration
parameters for the amount of time to kick reverse when changing
direction, and for how long, as well as how long to wait at
zero speed between throttle changes during direction change.