Implement recovery states
Added code to enter a recovery state if a cone is not seen when
a cone waypoint is achieved in AUTO mode. The recovery goes through
two states, each circling about a half circle. Removed unused code
and parameters. Added the ability to configure the mock cone
detector to avoid detecting cones.