Add recovery due to losing cone while driving to it
Added code to count the number of times the cone isn't detected
after the first detection. If the number of consecutive misses
is greater than a threshold, go into the recovery pattern. Added
configuration parameters for the maximum number of cone misses
before recovery. Fixed a problem in the mock cone detector where
the number of times a pattern position was used was one more
than the configured value. Fixed a bug where a nonexistent timer
was referenced.